requirements.txt
appdirs==1.4.3
asgiref==3.2.10
azure-core==1.8.1
azure-identity==1.4.0
azure-storage-blob==12.5.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.1
distlib==0.3.0
distro==1.4.0
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isodate==0.6.0
lockfile==0.12.2
msal==1.5.0
msal-extensions==0.2.2
msgpack==0.6.2
msrest==0.6.19
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
portalocker==1.7.1
progress==1.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
